refactor(slashCommands): improve file cohesion
