+++ src/commands/prefix/slashCommands/list-slash-commands.ts
@@ -1,8 +1,8 @@
 import { ApplicationCommand, EmbedBuilder } from "discord.js";
-import { SlashCommand } from "../../../types/utils";
 import djsRestHelper from "../../../utils/discordjs/slashCommands";
 import { listSlashCommands } from "../../../utils/slashCommandsListing";
 import { PrefixCommand } from "../prefixCommands.handler";
+import { SlashCommand } from "../../slash/slashCommands.handler";
 
 const renderSymbol = (value: boolean) => (value ? "  ✓   " : "  ✗   ");
