+++ src/commands/slash/ping.ts
@@ -1,7 +1,7 @@
 import PING_CONSTANTS from "../../constants/ping";
-import { SlashCommand } from "../../types/utils";
+import { SlashCommand } from "./slashCommands.handler";
 
-export default {
+const command: SlashCommand = {
   name: "ping",
   description: PING_CONSTANTS.description,
   type: "command",
@@ -12,4 +12,6 @@ export default {
   execute: async (_client, interaction) => {
     await interaction.reply(PING_CONSTANTS.pong);
   },
-} satisfies SlashCommand;
+};
+
+export default command;
