+++ src/types/common.d.ts
@@ -1,6 +1,6 @@
 import { PrefixCommand } from "../commands/prefix/prefixCommands.handler";
-import { SlashCommand } from "./utils";
 import { Collection } from "discord.js";
+import { SlashCommand } from "../commands/slash/slashCommands.handler";
 
 declare module "discord.js" {
   export interface Client {
